backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
motor==3.4.0
pydantic==2.7.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
